Move extension methods to .Internal and XML doc cleanup
