+++ benchmarks/MiniProfiler.Benchmarks/Benchmarks/SerializationBenchmarks.cs
@@ -1,7 +1,7 @@
 ﻿using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Attributes.Jobs;
 using StackExchange.Profiling;
-using StackExchange.Profiling.Helpers;
+using StackExchange.Profiling.Internal;
 
 namespace Benchmarks
 {
