+++ src/MiniProfiler.Shared/Internal/MiniProfilerExtensions.cs
@@ -1,5 +1,4 @@
-﻿using StackExchange.Profiling.Helpers;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
